requirements.txt
discord.py
pyowm
schedule